forked rashids front end, put it in utils/output_panel and modified it according to out can messages. Corrected length when writing from drivetrain to fix problem with showing speed

+++ project/source_code/drivetrain/src/vehicle.cpp
@@ -59,7 +59,7 @@ bool Vehicle::Run()
 
             payload[0] = static_cast<uint8_t>(this->engine->get_eng_sts());
             payload[1] = static_cast<uint8_t>(this->engine->get_eng_rpm()/37);
-            payload[2] = vehicle_speed;
+            payload[2] = static_cast<uint8_t>(vehicle_speed);
             CanBuffer::GetInstance().AddTx(&transmit_id, payload, &transmit_length);
             return_value = 1;
         }
